consult by topsis

+++ internal/domain/model/consult.go
@@ -4,7 +4,7 @@ import "topsis/handler/model"
 
 type ConsultResult struct {
 	Name       string  `json:"name"`
-	Similarity float32 `json:"similarity"`
+	Similarity float64 `json:"similarity"`
 }
 
 type MetadataScoreRating struct {
